requirements.txt
flask
python-dotenv
tensorflow
torch
transformers
